Add files via upload
The resume classifier(Resumate)using NLP for a dataset of 2500 rows and 21 categories is a data science project with the objective of automating the recruitment process for a company. The project involves multiple stages such as data collection, data cleaning, data preprocessing, feature engineering, model building, and model evaluation.

The first stage involves collecting resumes from various sources and labeling them based on their categories. This labeled data is then used for training and testing the model. The data cleaning stage involves removing irrelevant information, converting the text to lowercase, and removing stop words, punctuation, and special characters.

The next stage is data preprocessing, where the text data is tokenized, lemmatize, and tagged with part-of-speech information. These features are then used for feature engineering, where the most relevant features are selected for each category. The model is built using machine learning algorithms such as Naive Bayes, Support Vector Machines, Random Forest, Gradient Boosting, and Decision Tree and is trained using the labeled data.

The model's accuracy is evaluated using metrics such as precision, recall, and F1-score. The hyperparameters of the model are fine-tuned to optimize its performance, and the model is then deployed for use in the recruitment process.

The business insights that can be derived from the project include an automated and streamlined recruitment process, reduced recruitment costs, and improved quality of candidate screening. The model can also provide insights into hiring trends and patterns across different categories, such as the most sought-after skills and qualifications, the most common gaps in candidate skillsets, and the most effective channels for candidate sourcing.
Overall, the resume classifier using NLP for a dataset of 2500 rows and 21 categories is a valuable data science project that can help organizations optimize their recruitment process, reduce costs, and gain insights into hiring trends and patterns.